Move ASM wrappers to metacp.asm package.
